Koshucode.Baala.Base.Data.Dataset

Description

Dataset as a set of judges.

Dataset is like a bridge of judges and relations. We can get a relation from a dataset, that dataset is build from judges.

Documentation

data Dataset v

Dataset is a set of judges.

emptyDataset :: Dataset v

Dataset that has no judges

dataset :: [Judge v] -> Dataset v

Gather judges into a dataset

addJudges :: [Judge v] -> Dataset v -> Dataset v

Add judges to dataset.

selectRelation

Arguments

:: (Ord v, Nil v)

=> Dataset v
Dataset

-> Relsign
Relsign to select

-> [String]
List of term names

-> Rel v
Selected relation

Select relation from dataset. If a giving term is not in judges, Nil sign is used.