Koshucode.Baala.Base.Struct.Half.HalfRelmap

Relmap implementation

makeConsRelmap

Arguments

:: [RelmapImplement v]
Implementations of relmap operator

-> ConsRelmap v
Relmap constructors

Make half and full relmap constructors.

data ConsRelmap v

Half and full relmap constructor

Constructors

ConsRelmap

Fields

consHalf :: ConsHalfRelmap

consFull :: ConsFullRelmap v

data RelmapImplement v

Implementation of relmap operator. It consists of (1) operator name, (2) operand parser, (3) full constructor, and (4) usage of operator.

Constructors

RelmapImplement String RelmapParser (ConsFullRelmap v) [String]

type RelmapParser = [TokenTree] -> [(String, [TokenTree])]

Parser for operand of relational operator. This parsers docompose operand trees, and give a name to suboperand.

Half constructor

data HalfRelmap

Intermediate data that represents use of relmap operator. It consists of (1) operator name as String, (2) usage of operator as list of String, (3) its operands as list of RelmapOperand, and (4) subrelmaps in operands as list of HalfRelmap.

Constructors

HalfRelmap String [String] [(String, [TokenTree])] [HalfRelmap]

type ConsHalfRelmap

Arguments

= String
Operator name

-> [TokenTree]
Operand as source trees

-> HalfRelmap
Result half relmap with empty subs

First step of relmap construction, make HalfRelmap from use of relational operator.

consHalfRelmap :: ConsHalfRelmap -> [TokenTree] -> HalfRelmap

Construct half relmaps from source trees.

Full constructor

type ConsFullRelmap v

Arguments

= String
Operator name from HalfRelmap

-> [(String, [TokenTree])]
Operands from HalfRelmap

-> [Relmap v]
Subrelmaps computed by consFullRelmap

-> AbortOr (Relmap v)
Result relmap

Second step of relmap synthesis, make Relmap from contents of HalfRelmap.

consFullRelmap :: ConsFullRelmap v -> HalfRelmap -> AbortOr (Relmap v)

Construct (full) relmap.