Koshucode.Baala.Base.Struct.ModuleIO

Reading module

moduleRead

Arguments

:: StringValue v

=> Module v
Module that is same type to result module

-> String
Source text

-> AbortOr (Module v)
Result module from source text

moduleFile :: StringValue v => Module v -> FilePath -> IO (AbortOr (Module v))

Read module from file.

Running module

runModuleJudges

Arguments

:: Value v

=> Module v
Editing module

-> [Judge v]
Input judges

-> AbortOr [Judge v]
Output judges

Convert input judges into output judges by editing module. This function calls runAssert for list of assertions in module, and runAssert calls runRelmap for relmap in assertion.

runModuleString :: Value v => Module v -> String -> AbortOr String

Convert input data into output data by editing module.

runModuleIO :: Value v => Module v -> IO ()

Run editing module.