Koshucode.Baala.Base.Struct.Full.Relmap

Description

Data structures for relation-to-relation mappings

Relmap

data Relmap v

Relation-to-relation mapping. A Relmap is correspond to a pair of operator and its operands, that is a use of relational operator. Calculations for relations are done by runRelmap.

This picture represents calculation of mapping input relation to output relation.

   input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

   R1 -[ A ]- R2
   R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

   R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also Relmap of RelmapAppend A B.

   R1 -[ A ]--[ B ]- R3

Constructors

RelmapSource HalfRelmap String [String]
Relmap that retrieves a relation from a dataset

RelmapConst HalfRelmap String (Rel v)
Relmap that has a constant relation

RelmapAlias HalfRelmap (Relmap v)
Relmap that has equavalent relmap

RelmapCalc HalfRelmap String (RelmapSub v) [Relmap v]
Relmap that maps relations to a relation

RelmapAppend (Relmap v) (Relmap v)
Relmap that joints two relmaps

RelmapName String
Relmap reference

type RelmapSub v = [Rel v] -> Rel v -> Rel v

Function of relmap

type RelmapBin v = Relmap v -> Relmap v

Confluent operator, like binary operator

relmapSourceList :: Relmap v -> [Relmap v]

List of RelmapSource

relmapAppendList :: Relmap v -> [Relmap v]

Expand RelmapAppend to list of Relmap

calc :: HalfRelmap -> String -> RelmapSub v -> Relmap v

Make a non-confluent relmap

confl :: HalfRelmap -> String -> RelmapSub v -> [Relmap v] -> Relmap v

Make a confluent relmap

flow :: String -> RelmapFun v -> Relmap v

Make a flow relmap

Half

data HalfRelmap

Intermediate data that represents use of relational operator.

HalfRelmap is constructed from list of TokenTree, and (full) Relmap is constructed from HalfRelmap.

Constructors

HalfRelmap

Fields

halfUsage :: [String]

Usages description

halfOperator :: String

Operator name of relmap operation

halfOperand :: [Named [TokenTree]]

Operand of relmap operation

halfSubmap :: [HalfRelmap]

Subrelmaps in the operand

Run

runRelmap

Arguments

:: (Ord v, Nil v)

=> Dataset v
Judges read from source operator

-> Relmap v
Mapping from Rel to Rel

-> Rel v
Input relation

-> AbortOr (Rel v)
Output relation

Calculate Relmap for Rel.